Multi threaded file activity.
The two long running file actions (checking file info and actual backup) have been split to allow multi threaded processing.
This is because the CB Protection driver has a serious performance hit with files that have a large number of hard links.
The main thread generates a queue of items to work on and passes the queue (along with other parameters) to worker functions.