Add support to compress certain files instead of generating a binary delta.
This is to make future archiving of external backups easier. The daily backups will use the binary delta,
It makes very small delta files. For the weekly backups, I want to easily move some backups to an archival area.
This requires each weekly backup having a full copy of the mail data (otherwise I need to jump through hoops).
So the weeklies compress full mail data, then it is a simple date selection to choose which weeklies are archived.